Create new project header command
To reduce code duplication, base both the new project header and
the existing location header on a new base header.

Note that institution and location are combined in the project header
because some "institutions" such as "Personal" do not have locations.
(This means they must be combined in the base header as well.)