Change "Industry" to "City of Industry"
The first, simpler reason is that "City of Industry" is more common.

The second reason is that the place name in a mailing address is more useful
than the actual city, and USPS doesn't recognize "Industry" as a place name
for the ZIP Code.

Case in point: Yang-Ming International Corp. wasn't in La Puente at all;
it was actually in South San Jose Hills, but it used La Puente as
the place name.