Create "hierarchy of indentation"
The hierarchy of indentation (relative to left margin) is:
* Section headings: 0 in.
* Primary items: 1/4 in.
* Secondary items (experience points): 3/8 in.

Remove the symbol decoration on section headings.

Explicitly set indentation for list items (from left edge of bullet
to text) to 1/8 in.

For convenience, set the left page margin to 3/4 in. to align with
the primary indentation. To restore the "apparent" margin to 1/2 in.,
set the indentation for section headings to -1/2 in. and adjust
the centered top part similarly.